Use for_each to make the worker module conditional
Using for_earch with a boolean is a trick to allow to make inclusion of
an entire module conditional. A lot of the resources in the server
module used for populating workers were already conditional, but the
placement group wasn't, leading to a useless group being created.
Furthermore, just avoiding going through it if not needed is useful